Code/requirements.txt
torch==1.13.1
torchvision==0.12.0+cu113
Pillow==9.1.1
numpy==1.22.3
